Covitest Portal AI Project/requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.4
pycparser==2.21
python-google-places==1.4.2
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
SQLAlchemy==1.4.27
threadpoolctl==3.0.0
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
